+++ Elements/src/Geometry/Arc.cs
@@ -11,26 +11,68 @@ namespace Elements.Geometry
     /// <example>
     /// [!code-csharp[Main](../../Elements/test/ArcTests.cs?name=example)]
     /// </example>
-    public partial class Arc : Curve, IEquatable<Arc>
+    public partial class Arc : TrimmedCurve<Circle>, IEquatable<Arc>
     {
-        /// <summary>The center of the arc.</summary>
-        [JsonProperty("Center", Required = Required.AllowNull)]
-        public Vector3 Center { get; set; }
-
-        /// <summary>The radius of the arc.</summary>
-        [JsonProperty("Radius", Required = Required.Always)]
-        [System.ComponentModel.DataAnnotations.Range(0.0D, double.MaxValue)]
-        public double Radius { get; set; }
-
         /// <summary>The angle from 0.0, in degrees, at which the arc will start with respect to the positive X axis.</summary>
         [JsonProperty("StartAngle", Required = Required.Always)]
         [System.ComponentModel.DataAnnotations.Range(0.0D, 360.0D)]
-        public double StartAngle { get; set; }
+        public double StartAngle { get; protected set; }
 
         /// <summary>The angle from 0.0, in degrees, at which the arc will end with respect to the positive X axis.</summary>
         [JsonProperty("EndAngle", Required = Required.Always)]
         [System.ComponentModel.DataAnnotations.Range(0.0D, 360.0D)]
-        public double EndAngle { get; set; }
+        public double EndAngle { get; protected set; }
+
+        /// <summary>
+        /// The radius of the arc.
+        /// </summary>
+        public double Radius
+        {
+            get
+            {
+                return this.BasisCurve.Radius;
+            }
+        }
+
+        /// <summary>
+        /// The center of the arc.
+        /// </summary>
+        public Vector3 Center
+        {
+            get
+            {
+                return this.BasisCurve.Transform.Origin;
+            }
+        }
+
+        /// <summary>
+        /// The start point of the arc.
+        /// </summary>
+        [JsonIgnore]
+        public override Vector3 Start
+        {
+            get { return PointAt(this.Domain.Min); }
+        }
+
+        /// <summary>
+        /// The end point of the arc.
+        /// </summary>
+        [JsonIgnore]
+        public override Vector3 End
+        {
+            get { return PointAt(this.Domain.Max); }
+        }
+
+        /// <summary>
+        /// Create a circular arc.
+        /// </summary>
+        public Arc(double radius)
+        {
+            this.BasisCurve = new Circle();
+            this.StartAngle = 0;
+            this.EndAngle = 360;
+            this.Domain = new Domain1d(Units.DegreesToRadians(this.StartAngle), Units.DegreesToRadians(this.EndAngle));
+        }
 
         /// <summary>
         /// Create an arc.
@@ -42,11 +84,21 @@ namespace Elements.Geometry
         [JsonConstructor]
         public Arc(Vector3 @center, double @radius, double @startAngle, double @endAngle) : base()
         {
+            Validate(startAngle, endAngle, radius);
+            this.BasisCurve = new Circle(@center, @radius);
+            this.StartAngle = @startAngle;
+            this.EndAngle = @endAngle;
+            this.Domain = new Domain1d(Units.DegreesToRadians(@startAngle), Units.DegreesToRadians(@endAngle));
+        }
+
+        private void Validate(double startAngle, double endAngle, double radius)
+        {
             if (!Validator.DisableValidationOnConstruction)
             {
-                if (endAngle > 360.0 || startAngle > 360.00)
+                var span = Math.Abs(endAngle - startAngle);
+                if (span > 360)
                 {
-                    throw new ArgumentOutOfRangeException("The arc could not be created. The start and end angles must be greater than -360.0");
+                    throw new ArgumentOutOfRangeException($"The total span of the arc, {span} degrees, is greater than 360.0. The arc cannot be created.");
                 }
 
                 if (endAngle == startAngle)
@@ -59,26 +111,58 @@ namespace Elements.Geometry
                     throw new ArgumentOutOfRangeException($"The arc could not be created. The provided radius ({radius}) must be greater than 0.0.");
                 }
             }
+        }
 
-            this.Center = @center;
-            this.Radius = @radius;
-            this.StartAngle = @startAngle;
-            this.EndAngle = @endAngle;
+        /// <summary>
+        /// Create an arc.
+        /// Constructs a circular basis curve internally with a default transform.
+        /// </summary>
+        /// <param name="radius">The radius of the arc.</param>
+        /// <param name="startAngle">The CCW angle from 0.0, in degrees, at which the arc will start with respect to the positive X axis.</param>
+        /// <param name="endAngle">The CCW angle from 0.0, in degrees, at which the arc will end with respect to the positive X axis.</param>
+        public Arc(double radius, double startAngle, double endAngle)
+            : base()
+        {
+            Validate(startAngle, endAngle, radius);
+            this.BasisCurve = new Circle(radius);
+            this.StartAngle = startAngle;
+            this.EndAngle = endAngle;
+            this.Domain = new Domain1d(Units.DegreesToRadians(@startAngle), Units.DegreesToRadians(@endAngle));
         }
 
         /// <summary>
         /// Create an arc.
         /// </summary>
-        /// <param name="radius">The radius of the arc.</param>
-        /// <param name="startAngle">The angle from 0.0, in degrees, at which the arc will start with respect to the positive X axis.</param>
-        /// <param name="endAngle">The angle from 0.0, in degrees, at which the arc will end with respect to the positive X axis.</param>
-        public Arc(double radius, double startAngle, double endAngle)
-            : base()
+        /// <param name="circle">The circle on which this arc is based.</param>
+        /// <param name="startParameter">The parameter, from 0.0->2PI, of the start of the arc.</param>
+        /// <param name="endParameter">The parameter, from 0.0->2PI, of the end of the arc.</param>
+        public Arc(Circle circle, double startParameter, double endParameter)
         {
-            this.Center = Vector3.Origin;
-            this.Radius = radius;
-            this.StartAngle = startAngle;
-            this.EndAngle = endAngle;
+            Validate(Units.RadiansToDegrees(startParameter), Units.RadiansToDegrees(endParameter), circle.Radius);
+            this.BasisCurve = circle;
+            this.Domain = new Domain1d(startParameter, endParameter);
+            this.StartAngle = Units.RadiansToDegrees(this.Domain.Min);
+            this.EndAngle = Units.RadiansToDegrees(this.Domain.Max);
+        }
+
+        /// <summary>
+        /// Create an arc.
+        /// Constructs a circular basis curve internally with the provided transform.
+        /// </summary>
+        /// <param name="transform">The transform for the basis curve of the arc.</param>
+        /// <param name="radius">The radius of the arc.</param>
+        /// <param name="startParameter">The parameter, from 0.0->2PI, of the start of the arc.</param>
+        /// <param name="endParameter">The parameter, from 0.0->2PI, of the end of the arc.</param>
+        public Arc(Transform transform,
+                   double radius = 1.0,
+                   double startParameter = 0.0,
+                   double endParameter = Math.PI * 2)
+        {
+            Validate(Units.RadiansToDegrees(startParameter), Units.RadiansToDegrees(endParameter), radius);
+            this.BasisCurve = new Circle(transform, radius);
+            this.Domain = new Domain1d(startParameter, endParameter);
+            this.StartAngle = Units.RadiansToDegrees(this.Domain.Min);
+            this.EndAngle = Units.RadiansToDegrees(this.Domain.Max);
         }
 
         /// <summary>
@@ -86,57 +170,15 @@ namespace Elements.Geometry
         /// </summary>
         public override double Length()
         {
-            return 2 * Math.PI * this.Radius * (Math.Abs(this.EndAngle - this.StartAngle)) / 360.0;
+            return 2 * Math.PI * this.BasisCurve.Radius * (Math.Abs(this.EndAngle - this.StartAngle)) / 360.0;
         }
 
         /// <summary>
-        /// The start point of the arc.
+        /// The mid point of the line.
         /// </summary>
-        [JsonIgnore]
-        public Vector3 Start
+        public override Vector3 Mid()
         {
-            get { return PointAt(0.0); }
-        }
-
-        /// <summary>
-        /// The end point of the arc.
-        /// </summary>
-        [JsonIgnore]
-        public Vector3 End
-        {
-            get { return PointAt(1.0); }
-        }
-
-        /// <summary>
-        /// Return the point at parameter u on the arc.
-        /// </summary>
-        /// <param name="u">A parameter between 0.0 and 1.0.</param>
-        /// <returns>A Vector3 representing the point along the arc.</returns>
-        public override Vector3 PointAt(double u)
-        {
-            if (u > 1.0 + Vector3.EPSILON || u < 0.0 - Vector3.EPSILON)
-            {
-                throw new ArgumentOutOfRangeException($"The value provided for parameter u, {u}, must be between 0.0 and 1.0.");
-            }
-
-            var angle = this.StartAngle + (this.EndAngle - this.StartAngle) * u;
-            var theta = DegToRad(angle);
-            var x = this.Center.X + this.Radius * Math.Cos(theta);
-            var y = this.Center.Y + this.Radius * Math.Sin(theta);
-            return new Vector3(x, y);
-        }
-
-        /// <summary>
-        /// Return transform on the arc at parameter u.
-        /// </summary>
-        /// <param name="u">A parameter between 0.0 and 1.0 on the arc.</param>
-        /// <returns>A transform with its origin at u along the curve and its Z axis tangent to the curve.</returns>
-        public override Transform TransformAt(double u)
-        {
-            var p = PointAt(u);
-            var x = (p - this.Center).Unitized();
-            var y = Vector3.ZAxis;
-            return new Transform(p, x, x.Cross(y));
+            return PointAt(this.Domain.Min + this.Domain.Length / 2);
         }
 
         /// <summary>
@@ -144,12 +186,7 @@ namespace Elements.Geometry
         /// </summary>
         public Arc Reversed()
         {
-            return new Arc(this.Center, this.Radius, this.EndAngle, this.StartAngle);
-        }
-
-        private double DegToRad(double degrees)
-        {
-            return degrees * Math.PI / 180.0;
+            return new Arc(this.BasisCurve.Transform.Origin, this.BasisCurve.Radius, this.EndAngle, this.StartAngle);
         }
 
         /// <summary>
@@ -158,9 +195,9 @@ namespace Elements.Geometry
         /// <returns>A bounding box for this arc.</returns>
         public override BBox3 Bounds()
         {
-            var delta = new Vector3(this.Radius, this.Radius, this.Radius);
-            var min = new Vector3(this.Center - delta);
-            var max = new Vector3(this.Center + delta);
+            var delta = new Vector3(this.BasisCurve.Radius, this.BasisCurve.Radius, this.BasisCurve.Radius);
+            var min = new Vector3(this.BasisCurve.Transform.Origin - delta);
+            var max = new Vector3(this.BasisCurve.Transform.Origin + delta);
             return new BBox3(min, max);
         }
 
@@ -170,53 +207,49 @@ namespace Elements.Geometry
         /// <returns>The plane in which the arc lies.</returns>
         public Plane Plane()
         {
-            return new Plane(this.PointAt(0.0), this.PointAt(1.0), this.Center);
-        }
-
-        internal override double[] GetSampleParameters(double startSetback = 0.0, double endSetback = 0.0)
-        {
-            // Arc length calculations.
-            // var l = this.Length();
-            // var arcLength = l - startSetback - endSetback;
-            // al = (alpha * PI * r) / 180
-            // alpha = (al * 180)/(PI * r)
-            // var arcAngle = (arcLength * 360) / (2 * Math.PI * this.Radius);
-            // var parameterSpan = 1.0 - startSetback/l - endSetback/l;
-
-            // Parameter calculations.
-            var angleSpan = this.EndAngle - this.StartAngle;
-            var partialAngleSpan = Math.Abs(angleSpan - angleSpan * startSetback - angleSpan * endSetback);
-            var parameterSpan = 1.0 - 1.0 * startSetback - 1.0 * endSetback;
-
-            // Angle span: t
-            // d = 2 * r * sin(t/2)
-            var r = this.Radius;
-            var two_r = 2 * r;
-            var d = Math.Min(MinimumChordLength, two_r);
-            var t = 2 * Math.Asin(d / two_r);
-            var div = (int)Math.Ceiling((DegToRad(partialAngleSpan)) / t);
-
-            var parameters = new double[div + 1];
-            for (var i = 0; i <= div; i++)
-            {
-                var u = startSetback + i * (parameterSpan / div);
-                parameters[i] = u;
-            }
-            return parameters;
+            return BasisCurve.Transform.XY();
         }
 
         /// <summary>
-        /// A list of vertices describing the arc for rendering.
+        /// Get parameters to be used to find points along the curve for visualization.
         /// </summary>
-        internal override IList<Vector3> RenderVertices()
+        /// <param name="startSetbackDistance">An optional setback from the start of the curve.</param>
+        /// <param name="endSetbackDistance">An optional setback from the end of the curve.</param>
+        public override double[] GetSubdivisionParameters(double startSetbackDistance = 0.0,
+                                                       double endSetbackDistance = 0.0)
         {
-            var parameters = GetSampleParameters();
-            var vertices = new List<Vector3>();
-            foreach (var p in parameters)
+            var min = this.Domain.Min;
+            var max = this.Domain.Max;
+
+            var flip = max < min;
+
+            if (flip)
             {
-                vertices.Add(PointAt(p));
+                max = this.Domain.Min;
+                min = this.Domain.Max;
             }
-            return vertices;
+
+            var startParam = ParameterAtDistanceFromParameter(startSetbackDistance, min);
+            var endParam = ParameterAtDistanceFromParameter(this.Length() - endSetbackDistance, min);
+
+            // Parameter calculations.
+            var angleSpan = endParam - startParam;
+
+            // Angle span: t
+            // d = 2 * r * sin(t/2)
+            var r = this.BasisCurve.Radius;
+            var two_r = 2 * r;
+            var d = Math.Min(MinimumChordLength, two_r);
+            var t = 2 * Math.Asin(d / two_r);
+            var div = (int)Math.Ceiling(angleSpan / t);
+
+            var parameters = new double[div + 1];
+            var step = angleSpan / div;
+            for (var i = 0; i <= div; i++)
+            {
+                parameters[i] = startParam + i * step;
+            }
+            return parameters;
         }
 
         /// <summary>
@@ -230,7 +263,7 @@ namespace Elements.Geometry
             {
                 return false;
             }
-            return this.Center.Equals(other.Center) && this.StartAngle == other.StartAngle && this.EndAngle == other.EndAngle;
+            return this.BasisCurve.Transform.Origin.Equals(other.BasisCurve.Transform.Origin) && this.StartAngle == other.StartAngle && this.EndAngle == other.EndAngle;
         }
 
         /// <summary>
@@ -245,7 +278,7 @@ namespace Elements.Geometry
             {
                 newStart = newStart - 360.0;
             }
-            return new Arc(this.Center, this.Radius, newStart, newEnd);
+            return new Arc(this.BasisCurve.Transform.Origin, this.BasisCurve.Radius, newStart, newEnd);
         }
 
         /// <summary>
@@ -263,7 +296,76 @@ namespace Elements.Geometry
         /// <param name="transform">The transform to apply.</param>
         public Arc TransformedArc(Transform transform)
         {
-            return new Arc(transform.OfPoint(Center), Radius, StartAngle, EndAngle);
+            return new Arc(transform.OfPoint(this.BasisCurve.Transform.Origin), this.BasisCurve.Radius, StartAngle, EndAngle);
+        }
+
+        /// <summary>
+        /// Get the point at parameter u.
+        /// </summary>
+        /// <returns>The point at parameter u if us is within the trim, otherwise an exception is thrown.</returns>
+        public override Vector3 PointAt(double u)
+        {
+            if (!this.Domain.Includes(u, true))
+            {
+                throw new Exception($"The parameter {u} is not on the trimmed portion of the basis curve. The parameter must be between {Domain.Min} and {Domain.Max}.");
+            }
+            return this.BasisCurve.PointAt(u);
+        }
+
+        /// <summary>
+        /// Get the transform at parameter u.
+        /// </summary>
+        /// <returns>The transform at parameter u if us is within the trim, otherwise an exception is thrown.</returns>
+        public override Transform TransformAt(double u)
+        {
+            if (!this.Domain.Includes(u, true))
+            {
+                throw new Exception($"The parameter {u} is not on the trimmed portion of the basis curve. The parameter must be between {Domain.Min} and {Domain.Max}.");
+            }
+            return this.BasisCurve.TransformAt(u);
+        }
+
+        /// <summary>
+        /// Create a polyline through a set of points along the curve.
+        /// </summary>
+        /// <param name="divisions">The number of divisions of the curve.</param>
+        /// <returns>A polyline.</returns>
+        public override Polyline ToPolyline(int divisions = 10)
+        {
+            var pts = new List<Vector3>(divisions + 1);
+            var step = this.Domain.Length / divisions;
+            for (var t = this.Domain.Min; t < this.Domain.Max; t += step)
+            {
+                pts.Add(PointAt(t));
+            }
+
+            // We don't go all the way to the end parameter, and 
+            // add it here explicitly because rounding errors can
+            // cause small imprecision which accumulates to make
+            // the final parameter slightly more/less than the actual
+            // end parameter.
+            pts.Add(PointAt(this.Domain.Max));
+            return new Polyline(pts);
+        }
+
+        /// <summary>
+        /// Get the parameter at a distance from the start parameter along the curve.
+        /// </summary>
+        /// <param name="distance">The distance from the start parameter.</param>
+        /// <param name="start">The parameter from which to measure the distance.</param>
+        public override double ParameterAtDistanceFromParameter(double distance, double start)
+        {
+            if (!Domain.Includes(start, true))
+            {
+                throw new Exception($"The parameter {start} is not on the trimmed portion of the basis curve. The parameter must be between {Domain.Min} and {Domain.Max}.");
+            }
+
+            if (distance == 0.0)
+            {
+                return start;
+            }
+
+            return this.BasisCurve.ParameterAtDistanceFromParameter(distance, start);
         }
     }
 }