+++ Elements/src/Geometry/Contour.cs
@@ -13,23 +13,23 @@ namespace Elements.Geometry
     /// </example>
     public class Contour : IEnumerable<Curve>
     {
-        private List<Curve> _curves = new List<Curve>();
+        private List<BoundedCurve> _curves = new List<BoundedCurve>();
 
         /// <summary>
         /// Construct a contour.
         /// </summary>
         /// <param name="curves">A list of curves to create the contour.</param>
         /// <exception>Throws an ArgumentException when the provided curves are not contiguous.</exception>
-        public Contour(List<Curve> curves)
+        public Contour(List<BoundedCurve> curves)
         {
             for (var i = 0; i <= curves.Count - 1; i++)
             {
                 var a = curves[i];
                 var next = i == curves.Count - 1 ? 0 : i + 1;
                 var b = curves[next];
-                if (a.PointAt(1).IsAlmostEqualTo(b.PointAt(0)) ||
-                    a.PointAt(1).IsAlmostEqualTo(b.PointAt(1)) ||
-                    a.PointAt(0).IsAlmostEqualTo(b.PointAt(0)))
+                if (a.End.IsAlmostEqualTo(b.Start) ||
+                    a.End.IsAlmostEqualTo(b.End) ||
+                    a.Start.IsAlmostEqualTo(b.Start))
                 {
                     continue;
                 }